refactor: remove components from page-view
