+++ apps/web/src/features/cluster/components/ror-cli-button.tsx
@@ -4,7 +4,7 @@ import { getRorLoginViewItem } from '../utils/cluster'
 import type { ClusterListViewRowType } from '@ror/js-api-client'
 import { Copy } from 'lucide-react'
 
-export const RorCliButton = ({ cluster }: ClusterListViewRowType) => (
+export const RorCliButton = (cluster: ClusterListViewRowType) => (
   <Button
     variant='rorcli'
     onClick={(e) => {
